+++ pkg/agent/event_types_test.go
@@ -169,7 +169,8 @@ func (r *recordingVisitor) VisitMemoryLoaded(MemoryLoadedEvent)        { r.visit
 func (r *recordingVisitor) VisitMemoryConsolidated(MemoryConsolidatedEvent) {
 	r.visited = "memory_consolidated"
 }
-func (r *recordingVisitor) VisitUnknown(UnknownEvent) { r.visited = "unknown" }
+func (r *recordingVisitor) VisitRunCost(RunCostEvent)  { r.visited = "run_cost" }
+func (r *recordingVisitor) VisitUnknown(UnknownEvent)  { r.visited = "unknown" }
 
 func TestVisit_DispatchesToMatchingMethod(t *testing.T) {
 	cases := []struct {